cnc/pyproject.toml
[tool.poetry]
name = "cnc"
version = "0.1.0"
description = ""
authors = ["leadpogrommer"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^14.1"
pillow = "^11.0.0"
asyncinotify = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
